Login backend (behaves weirdly ATM, see descr)
Technically, the server returns 302 (Found) with a real account
specified, which means it accepts the authorization request. However:
* the debugging showed that there are two ongoing requests (one of them
is IINM a redirect to the homepage), which is completely unnecessary
* due to the reason above, it isn't possible to determine whether the
auth was successful or not (see AsynchronousPostRequest code for
implementation)
* the personalized features (e. g. filters) do not work with a real
account specified. either the cookie is not being saved, or something
very shady is going on there