<-- REVERTED bottom bar's shared ScrollView
It turns out that using a shared ScrollView across all tabs:
1) hurts performance and memory consumption, as the content of the said
tabs does not get recycled when not visible
2) requires hacky workarounds to resize the content
3) is prone to a handful of bugs related to fling scrolling and
interaction with the content behind it (ImageView, for example, requires
a custom touch event listener).

This commit effectively removes the component while retaining other
changes made. Note that the bottom bar is still not behaving as it
should.